apps/auth-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pydantic[email]
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
grpcio
grpcio-tools